+++ src/Resources/Wacc/ReturnOrders/ReturnOrders.ts
@@ -22,7 +22,7 @@ export default class ReturnOrders extends AbstractResource {
     });
 
     public complete: Function = this.createMethodFromPartialSpec({
-        method: 'POST',
+        method: 'PUT',
         path: '/{id}/complete',
     });
 }