+++ src/__tests__/Client.spec.ts
@@ -17,6 +17,7 @@ import UsersBaskets from '../Resources/Users/Baskets';
 import UsersGifts from '../Resources/Users/Gifts';
 import UsersToybox from '../Resources/Users/Toybox';
 import UsersSubscriptions from '../Resources/Users/Subscriptions';
+import UsersProductLists from '../Resources/Users/ProductLists';
 import Gift from '../Resources/Gift/Gift';
 import Contact from '../Resources/Contact/Contact';
 
@@ -54,6 +55,7 @@ describe('Client.ts', () => {
         expect(client.users.gifts).toBeInstanceOf(UsersGifts);
         expect(client.users.subscriptions).toBeInstanceOf(UsersSubscriptions);
         expect(client.users.toybox).toBeInstanceOf(UsersToybox);
+        expect(client.users.productLists).toBeInstanceOf(UsersProductLists);
         expect(client.gift).toBeInstanceOf(Gift);
         expect(client.contact).toBeInstanceOf(Contact);
     });